Tries moving the weight and dimensions input boxes to the modal.
